requirement.txt
# LangChain core
langchain==0.3.27
langchain_community==0.2.1
langchain_core==0.3.68
langchain-openai==0.3.27
langchain-hub==0.1.14

# Google Gemini
langchain-google-genai==1.0.1
google-generativeai==0.5.4
google-api-core==2.18.0  # Needed for proper exception handling

# Streamlit app
streamlit==1.46.1

# Tooling (e.g., DuckDuckGo search)
duckduckgo-search==5.2.2

# Optional for compatibility/logging
tqdm>=4.66.0